Use tokenize_str() in lindera-cli
