+++ lindera-cli/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "lindera-cli"
-version = "0.15.0"
+version = "0.15.1"
 edition = "2021"
-description = "A morphological analysis tool."
+description = "A morphological analysis command line interface."
 documentation = "https://docs.rs/lindera-cli"
 homepage = "https://github.com/lindera-morphology/lindera"
 repository = "https://github.com/lindera-morphology/lindera"
@@ -24,6 +24,7 @@ compress = ["lindera/compress"]
 [dependencies]
 anyhow = "1.0.58"
 clap = { version = "3.2.8", features = ["derive", "cargo"] }
+serde_json = "1.0.82"
 
 lindera = { version = "0.15.0", path = "../lindera" }
