+++ lindera/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "lindera"
-version = "0.15.0"
+version = "0.15.1"
 edition = "2021"
 description = "A morphological analysis library."
 documentation = "https://docs.rs/lindera"
